refactor(iroh): split client code into individual modds
